Change ticcmd_t's itemID from int8_t to uint16_t
In the ticcmd_t it is stored as 1 higher than the actual itemID.
Fixed a couple faulty error checks.

+++ source/d_ticcmd.h
@@ -69,7 +69,7 @@ struct ticcmd_t
    byte    chatchar;
    byte    buttons;
    byte    actions;
-   int8_t  itemID;        // MaxW: ID of used inventory item
+   uint16_t itemID;     // MaxW: ID of used inventory item (+ 1)
 };
 
 #if defined(_MSC_VER) || defined(__GNUC__)
